Share Python updater tests between CPU and GPU.
Prepare to remove the `gpu_hist` parameter.
