+++ python-package/xgboost/testing/updater.py
@@ -2,7 +2,8 @@
 
 import json
 from functools import partial, update_wrapper
-from typing import Any, Dict, List, Union, overload
+from string import ascii_lowercase
+from typing import Any, Dict, List, Literal, TypeAlias, Union, overload
 
 import numpy as np
 import pytest
@@ -13,6 +14,8 @@ from xgboost.data import is_pd_cat_dtype
 
 from ..core import DataIter
 from .data_iter import CatIter
+
+Device: TypeAlias = Literal["cpu", "cuda"]
 
 
 @overload
@@ -572,6 +575,155 @@ def check_categorical_missing(  # pylint: disable=too-many-arguments
 
     # Test with partition-based split
     run(USE_PART)
+
+
+def run_max_cat(tree_method: str, device: Device) -> None:
+    """Test data with size smaller than number of categories."""
+    import pandas as pd
+
+    rng = np.random.default_rng(0)
+    n_cat = 100
+    n = 5
+
+    X = pd.Series(
+        ["".join(rng.choice(list(ascii_lowercase), size=3)) for i in range(n_cat)],
+        dtype="category",
+    )[:n].to_frame()
+
+    reg = xgb.XGBRegressor(
+        enable_categorical=True,
+        tree_method=tree_method,
+        device=device,
+        n_estimators=10,
+    )
+    y = pd.Series(range(n))
+    reg.fit(X=X, y=y, eval_set=[(X, y)])
+    assert tm.non_increasing(reg.evals_result()["validation_0"]["rmse"])
+
+
+def run_invalid_category(tree_method: str, device: Device) -> None:
+    """Test with invalid categorical inputs."""
+    rng = np.random.default_rng()
+    # too large
+    X = rng.integers(low=0, high=4, size=1000).reshape(100, 10)
+    y = rng.normal(loc=0, scale=1, size=100)
+    X[13, 7] = np.iinfo(np.int32).max + 1
+
+    # Check is performed during sketching.
+    Xy = xgb.DMatrix(X, y, feature_types=["c"] * 10)
+    with pytest.raises(ValueError):
+        xgb.train({"tree_method": tree_method, "device": device}, Xy)
+
+    X[13, 7] = 16777216
+    Xy = xgb.DMatrix(X, y, feature_types=["c"] * 10)
+    with pytest.raises(ValueError):
+        xgb.train({"tree_method": tree_method, "device": device}, Xy)
+
+    # mixed positive and negative values
+    X = rng.normal(loc=0, scale=1, size=1000).reshape(100, 10)  # type: ignore
+    y = rng.normal(loc=0, scale=1, size=100)
+
+    Xy = xgb.DMatrix(X, y, feature_types=["c"] * 10)
+    with pytest.raises(ValueError):
+        xgb.train({"tree_method": tree_method, "device": device}, Xy)
+
+    if device == "cuda":
+        import cupy as cp
+
+        X, y = cp.array(X), cp.array(y)
+        with pytest.raises(ValueError):
+            Xy = xgb.QuantileDMatrix(X, y, feature_types=["c"] * 10)
+
+
+def run_adaptive(tree_method: str, weighted: bool, device: Device) -> None:
+    """Test for adaptive trees."""
+    rng = np.random.RandomState(1994)
+    from sklearn.datasets import make_regression
+    from sklearn.utils import stats
+
+    n_samples = 256
+    X, y = make_regression(  # pylint: disable=unbalanced-tuple-unpacking
+        n_samples, 16, random_state=rng
+    )
+    if weighted:
+        w = rng.normal(size=n_samples)
+        w -= w.min()
+        Xy = xgb.DMatrix(X, y, weight=w)
+        base_score = stats._weighted_percentile(  # pylint: disable=protected-access
+            y, w, percentile=50
+        )
+    else:
+        Xy = xgb.DMatrix(X, y)
+        base_score = np.median(y)
+
+    booster_0 = xgb.train(
+        {
+            "tree_method": tree_method,
+            "base_score": base_score,
+            "objective": "reg:absoluteerror",
+            "device": device,
+        },
+        Xy,
+        num_boost_round=1,
+    )
+    booster_1 = xgb.train(
+        {
+            "tree_method": tree_method,
+            "objective": "reg:absoluteerror",
+            "device": device,
+        },
+        Xy,
+        num_boost_round=1,
+    )
+    config_0 = json.loads(booster_0.save_config())
+    config_1 = json.loads(booster_1.save_config())
+
+    def get_score(config: Dict) -> float:
+        return float(config["learner"]["learner_model_param"]["base_score"])
+
+    assert get_score(config_0) == get_score(config_1)
+
+    with pytest.warns(Warning, match="Model format is default to UBJSON"):
+        raw_booster = booster_1.save_raw(raw_format="deprecated")
+    booster_2 = xgb.Booster(model_file=raw_booster)
+    config_2 = json.loads(booster_2.save_config())
+    assert get_score(config_1) == get_score(config_2)
+
+    raw_booster = booster_1.save_raw(raw_format="ubj")
+    booster_2 = xgb.Booster(model_file=raw_booster)
+    config_2 = json.loads(booster_2.save_config())
+    assert get_score(config_1) == get_score(config_2)
+
+    booster_0 = xgb.train(
+        {
+            "tree_method": tree_method,
+            "base_score": base_score + 1.0,
+            "objective": "reg:absoluteerror",
+            "device": device,
+        },
+        Xy,
+        num_boost_round=1,
+    )
+    config_0 = json.loads(booster_0.save_config())
+    np.testing.assert_allclose(get_score(config_0), get_score(config_1) + 1)
+
+    evals_result: Dict[str, Dict[str, list]] = {}
+    xgb.train(
+        {
+            "tree_method": tree_method,
+            "device": device,
+            "objective": "reg:absoluteerror",
+            "subsample": 0.8,
+            "eta": 1.0,
+        },
+        Xy,
+        num_boost_round=10,
+        evals=[(Xy, "Train")],
+        evals_result=evals_result,
+    )
+    mae = evals_result["Train"]["mae"]
+    assert mae[-1] < 20.0
+    assert tm.non_increasing(mae)
 
 
 def train_result(
